refactor(windows): split up IPC into sub-modules
