+++ rust/windows-client/src-tauri/src/client/debug_commands.rs
@@ -11,7 +11,7 @@ pub enum Cmd {
     NetworkChanges,
     TestIpc {
         #[command(subcommand)]
-        cmd: Option<client::ipc::Subcommand>,
+        cmd: Option<client::ipc::multi_process_tests::Subcommand>,
     },
     Wintun,
 }
@@ -21,7 +21,7 @@ pub fn run(cmd: Cmd) -> Result<()> {
         Cmd::Crash => crash(),
         Cmd::Hostname => hostname(),
         Cmd::NetworkChanges => client::network_changes::run_debug(),
-        Cmd::TestIpc { cmd } => client::ipc::test_subcommand(cmd),
+        Cmd::TestIpc { cmd } => client::ipc::multi_process_tests::run(cmd),
         Cmd::Wintun => wintun(),
     }
 }
